break out ProviderSerializer
